+++ hqporner_api/api.py
@@ -1,71 +1,68 @@
 import os
 import string
-from functools import cached_property, wraps
+from functools import cached_property
 from random import choice
-from typing import Any, Generator, Callable
+from typing import Generator
+from bs4 import BeautifulSoup
 
 from hqporner_api.modules.locals import *
 from hqporner_api.modules.errors import *
 from hqporner_api.modules.functions import *
 from hqporner_api.modules.progress_bars import *
 
-from base_api.base import Core
 
+class Checks:
+    """
+    Does the same as the decorators, but decorators are not good for IDEs because they get confused.
+    So I moved them here.
+    """
 
-def validate_url(func: Callable):
-    @wraps(func)
-    def wrapper(self, url, *args, **kwargs) -> Any:
-        if check_url(url):
-            return func(self, url, *args, **kwargs)
+    @classmethod
+    def check_url(cls, url: str):
+        match = PATTERN_CHECK_URL.match(url)
+        if match:
+            return url
+
         else:
             raise InvalidURL
-    return wrapper
 
+    @classmethod
+    def check_actress(cls, actress: str):
+        if actress.startswith("https://"):
+            match = PATTERN_CHECK_URL_ACTRESS.match(actress)
+            if match:
+                name_extraction = re.compile(r'https://hqporner.com/actress/(.+)')
+                name = name_extraction.search(actress).group(1)
+                return name
 
-def validate_actress(func: Callable):
-    @wraps(func)
-    def wrapper(self, actress,  *args, **kwargs) -> Callable:
-        if check_actress(actress):
-            return func(self, actress, *args, **kwargs)
-
-        else:
-            raise InvalidActress
-
-    return wrapper
-
-
-def validate_category(func: Callable):
-    @wraps(func)
-    def wrapper(self, category: Category, *args,  **kwargs):
-        if check_category(category):
-            return func(self, category, *args, **kwargs)
+            else:
+                raise InvalidActress
 
         else:
-            raise InvalidCategory
+            return actress
 
-    return wrapper
+            # I assume that if it's not a URL, the user was smart enough to enter just the name lol
 
 
 class Video:
-    """
-    Creates a Video object, which can be used to retrieve information of a HQPorner video and download it.
-    """
-    @validate_url
     def __init__(self, url):
-        """
-        Initialize a new Video instance.
+        self.url = Checks().check_url(url)
+        self.html_content = requests.get(url=self.url, headers=headers).content.decode("utf-8")
 
-        This method takes a URL, validates it, and then fetches the HTML content of the web page at that URL.
+    @classmethod
+    def fix_quality(cls, quality):
+        if isinstance(quality, Quality):
+            return quality
 
-        Args:
-            url (str): A URL string pointing to a video. The URL should be valid and accessible.
+        else:
+            if str(quality) == "best":
+                return Quality.BEST
 
-        Attributes:
-            url (str): The URL of the video.
-            html_content (str): The HTML content of the web page at the provided URL.
-        """
-        self.url = url
-        self.html_content = Core().get_content(url=url, headers=headers).decode("utf-8")
+            elif str(quality) == "half":
+                return Quality.HALF
+
+            elif str(quality) == "worst":
+                return Quality.WORST
 
     @cached_property
     def title(self) -> str:
@@ -154,7 +151,7 @@ class Video:
         :param callback:
         :return:
         """
-        quality = Core().fix_quality(quality)
+        quality = self.fix_quality(quality)
 
         cdn_urls = self.direct_download_urls
         quals = self.video_qualities
@@ -162,7 +159,7 @@ class Video:
 
         # Define the quality map
         if not quals:
-            raise NotAvailable
+            raise NotAvailable("This video can't be downloaded, because it uses an older HTML player.")
 
         quality_map = {
             Quality.BEST: max(quals, key=lambda x: int(x)),
@@ -172,7 +169,6 @@ class Video:
 
         selected_quality = quality_map[quality]
         download_url = f"https://{quality_url_map[selected_quality]}"
-
         response = requests.get(download_url, stream=True)
         file_size = int(response.headers.get('content-length', 0))
 
@@ -206,32 +202,64 @@ class Video:
         cleaned_title = ''.join([char for char in title if char in string.printable and char not in illegal_chars])
         return cleaned_title
 
+    def get_thumbnails(self):
+        """
+        Note: This function is very bad optimized, but there's no other way. This is also the reason why it's not cached
+        The first item in the index is the base thumbnail.
+        """
+
+        id_from_url_pattern = re.compile("hqporner.com/hdporn/(.*?)-")
+        id = id_from_url_pattern.search(self.url).group(1)
+        title = self.title
+        urls = []
+        scripts_under_divs = []
+        script = None
+
+        query = title.replace(" ", "+")
+        html_content = requests.get(url=f"{root_url}/?q={query}").content.decode("utf-8")
+        soup = BeautifulSoup(html_content, 'lxml')
+        divs = soup.find_all('div', class_='row')
+
+        for div in divs:
+            scripts = div.find_all('script')
+            scripts_under_divs.extend(scripts)
+
+        pattern = re.compile(r'"(//[^"]+)"')
+        for script in scripts_under_divs:
+            if f"preload_{id}" in script.text:
+                script = script.text
+                break
+
+        urls_ = pattern.findall(script)
+        main_thumbnail = urls_[0].replace("_1.jpg", "_main.jpg")
+        urls.append("https:" + main_thumbnail)
+        for url in urls_:
+            urls.append("https:" + url)
+
+        return urls
+
 
 class Client:
 
-    @validate_url
-    def get_video(self, url: str) -> Video:
+    @classmethod
+    def get_video(cls, url: str) -> Video:
         """
         :param url:
         :return: Video object
         """
         return Video(url)
 
-    @validate_actress
-    def get_videos_by_actress(self, name: str, pages: int = 5) -> Generator[Video, None, None]:
+    @classmethod
+    def get_videos_by_actress(cls, name: str, pages: int = 5) -> Generator[Video, None, None]:
         """
         :param name:
         :param pages: int: one page contains 46 videos
         :return:
         """
-        match = re.search("hqporner.com/actress/(.+)", name)
-        if match:
-            name = match.group(1)
-
-        name = name.replace(" ", "-")
+        name = Checks().check_actress(name)
         for page in range(1, int(pages + 1)):
             final_url = f"{root_url_actress}{name}/{page}"
-            html_content = Core().get_content(final_url).decode("utf-8")
+            html_content = requests.get(final_url, headers=headers).content.decode("utf-8")
             if not check_for_page(html_content):
                 break
 
@@ -241,15 +269,15 @@ class Client:
                 if PATTERN_CHECK_URL.match(url):
                     yield Video(url)
 
-    @validate_category
-    def get_videos_by_category(self, category: Category, pages=5) -> Generator[Video, None, None]:
+    @classmethod
+    def get_videos_by_category(cls, category: Category, pages=5) -> Generator[Video, None, None]:
         """
         :param category: Category: The video category
         :param pages: int: one page contains 46 videos
         :return:
         """
         for page in range(1, int(pages + 1)):
-            html_content = Core().get_content(f"{root_url_category}{category}/{page}").decode("utf-8")
+            html_content = requests.get(url=f"{root_url_category}{category}/{page}").content.decode("utf-8")
             if not check_for_page(html_content):
                 break
 
@@ -266,14 +294,14 @@ class Client:
         :return:
         """
         query = query.replace(" ", "+")
-        html_content = Core().get_content(f"{root_url}/?q={query}").decode("utf-8")
+        html_content = requests.get(url=f"{root_url}/?q={query}").content.decode("utf-8")
         match = PATTERN_CANT_FIND.search(html_content)
         if "Sorry" in match.group(1).strip():
             raise NoVideosFound
 
         else:
             for page in range(1, int(pages + 1)):
-                html_content = Core().get_content(f"{root_url}/?q={query}&p={page}").decode("utf-8")
+                html_content = requests.get(url=f"{root_url}/?q={query}&p={page}").content.decode("utf-8")
                 if not check_for_page(html_content):
                     break
 
@@ -291,10 +319,10 @@ class Client:
         """
         for page in range(1, int(pages + 1)):
             if sort_by == "all_time":
-                html_content = Core().get_content(f"{root_url_top}{page}", headers=headers).decode("utf-8")
+                html_content = requests.get(f"{root_url_top}{page}", headers=headers).content.decode("utf-8")
 
             else:
-                html_content = Core().get_content(f"{root_url_top}{sort_by}/{page}", headers=headers).decode("utf-8")
+                html_content = requests.get(f"{root_url_top}{sort_by}/{page}", headers=headers).content.decode("utf-8")
 
             if not check_for_page(html_content):
                 break
@@ -309,7 +337,7 @@ class Client:
         """
         :return: list: Returns all categories of HQporner as a list of strings
         """
-        html_content = Core().get_content("https://hqporner.com/categories", headers=headers).decode("utf-8")
+        html_content = requests.get("https://hqporner.com/categories", headers=headers).content.decode("utf-8")
         categories = PATTERN_ALL_CATEGORIES.findall(html_content)
         return categories
 
@@ -318,7 +346,7 @@ class Client:
         """
         :return: Video object (random video from HQPorner)
         """
-        html_content = Core().get_content(root_random, headers=headers).decode("utf-8")
+        html_content = requests.get(root_random, headers=headers).content.decode("utf-8")
         videos = PATTERN_VIDEOS_ON_SITE.findall(html_content)
         video = choice(videos) # The random-porn from HQPorner returns 3 videos, so we pick one of them
         return Video(f"{root_url}hdporn/{video}")
@@ -330,7 +358,7 @@ class Client:
         :return:
         """
         for page in range(1, int(pages + 1)):
-            html_content = Core().get_content(url=f"{root_brazzers}/{page}").decode("utf-8")
+            html_content = requests.get(url=f"{root_brazzers}/{page}").content.decode("utf-8")
             if not check_for_page(html_content):
                 break